refactor: create module folders
