+++ lib/app.dart
@@ -2,9 +2,9 @@ import 'package:crypto_converter/core/colors.dart';
 import 'package:crypto_converter/core/icons.dart';
 import 'package:crypto_converter/core/routes.dart';
 import 'package:crypto_converter/core/strings.dart';
-import 'package:crypto_converter/presenter/converter/cubits/converter_cubit.dart';
-import 'package:crypto_converter/presenter/converter/views/conversion_view.dart';
-import 'package:crypto_converter/presenter/settings/views/settings_view.dart';
+import 'package:crypto_converter/modules/converter/cubits/converter_cubit.dart';
+import 'package:crypto_converter/modules/converter/views/conversion_view.dart';
+import 'package:crypto_converter/modules/settings/views/settings_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
