+++ lib/modules/coin_api/repository/coin_repository.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
-import 'package:crypto_converter/infrastructure/coin_api/api/coin_api_client.dart';
-import 'package:crypto_converter/infrastructure/coin_api/models/exchange_rate.dart';
-import 'package:crypto_converter/infrastructure/coin_api/models/exchange_rate_data.dart';
+import 'package:crypto_converter/modules/coin_api/repository/api/coin_api_client.dart';
+import 'package:crypto_converter/modules/coin_api/repository/models/exchange_rate.dart';
+import 'package:crypto_converter/modules/coin_api/repository/models/exchange_rate_data.dart';
 
 class CoinFailure implements Exception {}
