+++ lib/modules/coin_api/repository/models/cryptocurrency.dart
