+++ lib/modules/coin_api/repository/models/exchange_rate.dart
