+++ lib/modules/coin_api/repository/models/exchange_rate_data.dart
