+++ lib/modules/converter/models/currency.dart
