+++ lib/modules/converter/views/conversion_view.dart
@@ -1,8 +1,8 @@
 import 'package:crypto_converter/core/strings.dart';
-import 'package:crypto_converter/domain/converter/models/currency.dart';
-import 'package:crypto_converter/domain/converter/models/exchange_rate.dart';
-import 'package:crypto_converter/presenter/converter/cubits/converter_cubit.dart';
-import 'package:crypto_converter/presenter/converter/widgets/crypto_card_widget.dart';
+import 'package:crypto_converter/modules/converter/models/currency.dart';
+import 'package:crypto_converter/modules/converter/models/exchange_rate.dart';
+import 'package:crypto_converter/modules/converter/cubits/converter_cubit.dart';
+import 'package:crypto_converter/modules/converter/widgets/crypto_card_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:pull_to_refresh_flutter3/pull_to_refresh_flutter3.dart';
