+++ test/app_test.dart
@@ -1,6 +1,6 @@
 import 'package:crypto_converter/core/icons.dart';
-import 'package:crypto_converter/presenter/converter/views/conversion_view.dart';
-import 'package:crypto_converter/presenter/settings/views/settings_view.dart';
+import 'package:crypto_converter/modules/converter/views/conversion_view.dart';
+import 'package:crypto_converter/modules/settings/views/settings_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:crypto_converter/app.dart';
