+++ test/mocks/classes.dart
@@ -1,6 +1,6 @@
 import 'package:bloc_test/bloc_test.dart';
-import 'package:crypto_converter/domain/converter/usecases/list/list_exchange_rate_usecase.dart';
-import 'package:crypto_converter/presenter/converter/cubits/converter_cubit.dart';
+import 'package:crypto_converter/modules/converter/usecases/list_exchange_rate/list_exchange_rate_usecase.dart';
+import 'package:crypto_converter/modules/converter/cubits/converter_cubit.dart';
 import 'package:hydrated_bloc/hydrated_bloc.dart';
 import 'package:mocktail/mocktail.dart';
