+++ test/mocks/setup.dart
@@ -1,4 +1,4 @@
-import 'package:crypto_converter/domain/converter/usecases/list/list_exchange_rate_dto.dart';
+import 'package:crypto_converter/modules/converter/usecases/list_exchange_rate/list_exchange_rate_dto.dart';
 import 'package:mocktail/mocktail.dart';
 
 import 'storage.dart';
